requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
python-dotenv
requests
huggingface_hub
